Add files via upload
Though the year of release is 2025, this file was made as a part of a school coding assignment in late 2023. The goal of said assignment was to put into practice everything we had learned about HTML/CSS and web-development. We were free to choose the type of website, with the requirements that the website should be made for a real or hypothetical company and that it should include a base amount of web pages. 
Since it was made for a Dutch IT class, all the front-end is also in Dutch.
This file holds significance as it was my first project in web-development.